+++ uefi/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Changed
 - Changed ordering of `proto::pci::PciIoAddress` to (bus -> dev -> fun -> reg -> ext_reg).
+- Return request with status as error data object for `proto::ata::pass_thru::AtaDevice`.
 
 # uefi - v0.36.1 (2025-11-05)